refector: auth 로직 type 통합
