+++ src/entity/signin/index.ts
@@ -1,3 +1,2 @@
-export * from './model/signinState';
 export * from './model/useSigninSelectors';
 export * from './lib/getStep';
