+++ src/view/signin/ui/SigninPage/index.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { useCurrentStep } from '~/entity/signin/model/useSigninSelectors';
 import { NicknameStep, PasswordStep } from '@/widget/signin';
-import type { SigninState } from '~/entity/signin/model/signinState';
+import type { SigninState } from '~/entity/signup/model/authState';
 
 const STEP_COMPONENTS: Record<SigninState['currentStep'], React.ComponentType> = {
   nickname: NicknameStep,
